requirements.txt
matplotlib
Pillow
numpy
imageio